remove scia rBuffer. data goes directuly to sm
I decided to remove completely the cBuffer for received data from serial
scia 'cose I'd parse byte by byte at sm reactive proccess instead of
pooling the cBuffer data length. It consume less ram, 'cose I use the
cBuffer from the events and it's more quikcly and reactive. I use a
parser sm to parse each byte. The [enter] doesn't have meaning in these
scheme, but backspace yes but it doesnt work by the moment...

+++ src/wdog.c
@@ -32,7 +32,7 @@ uint32_t readUptimeDays(void)
    return ((uptime/10)/3660)/24;
 }
 
-uint16_t Cmd_Uptime(int argc, char *argv[])
+uint16_t Cmd_Uptime(uint16_t argc, char *argv[])
 {
    sciPrintf("%d days %02d:%02d:%02d.%1d \r\n",
                   readUptimeDays  ( ),
